Created an autonomous program with odometry
Created "AutoBase," which will allow for the testing of odometry  to see how well it works, and act as a reference for where the robot will move to during a full autonomous program. The program will keep track of the robot's x position, y position, and rotation relative to the field, so that a precise autonomous program can be built around it. Currently, the program will also output the values of the robot's x, y, and theta relative to the field.

This program is still theoretical. It will need to be tested.